split utils.py in mcmc.py and utils.py
